+++ apps/backend/src/libs/modules/database/libs/enums/database-table-name.enum.ts
@@ -8,6 +8,7 @@ const DatabaseTableName = {
 	FRIENDS: "friends",
 	MIGRATIONS: "migrations",
 	NOTIFICATIONS: "notifications",
+	SECTION_STATUSES: "section_statuses",
 	USER_DETAILS: "user_details",
 	USERS: "users",
 	VENDORS: "vendors",