+++ apps/backend/src/libs/modules/server-application/server-application.ts
@@ -9,6 +9,7 @@ import { courseSectionController } from "~/modules/course-sections/course-sectio
 import { courseController } from "~/modules/courses/courses.js";
 import { fileController } from "~/modules/files/files.js";
 import { friendController } from "~/modules/friends/friends.js";
+import { sectionStatusController } from "~/modules/section-statuses/section-statuses.js";
 import { userCourseController } from "~/modules/user-courses/user-courses.js";
 import { userNotificationController } from "~/modules/user-notifications/user-notifications.js";
 import { userController, userService } from "~/modules/users/users.js";
@@ -31,6 +32,7 @@ const apiV1 = new BaseServerApplicationApi(
 	...userController.routes,
 	...userNotificationController.routes,
 	...courseSectionController.routes,
+	...sectionStatusController.routes,
 );
 const serverApplication = new BaseServerApplication({
 	apis: [apiV1],