+++ packages/shared/src/modules/section-statuses/libs/validation-schemas/section-status-get-all-query.validation-schema.ts
@@ -1,0 +1,10 @@
+import { z } from "zod";
+
+const sectionStatusGetAllQuery = z
+	.object({
+		courseId: z.coerce.number(),
+		userId: z.coerce.number(),
+	})
+	.required();
+
+export { sectionStatusGetAllQuery };